# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ProfileInput {
  /**
     * @minLength 3
     * @maxLength 20
     * @pattern ^[a-z0-9_]+$
     */
  username: string;
  /**
     * @minLength 1
     * @maxLength 32
     */
  displayName: string;
}

export interface CommunityProfile {
  userId: string;
  username: string;
  displayName: string;
  score: number;
  rank: string;
  rankLevel: number;
  friendCount: number;
  /** @nullable */
  nextRankScore: number | null;
  sharingEnabled: boolean;
}

export interface SharingPreferenceInput {
  enabled: boolean;
}

export interface SharingPreferenceResult {
  enabled: boolean;
}

export interface WorkoutSessionInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  clientSessionId: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  workoutId: string;
  /**
     * @minimum 1
     * @maximum 100
     */
  sets: number;
  /**
     * @minimum 1
     * @maximum 240
     */
  minutes: number;
  completedAt: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  localDay: string;
}

export interface WorkoutSyncInput {
  /**
     * @minItems 1
     * @maxItems 250
     */
  sessions: WorkoutSessionInput[];
  /**
     * @minLength 1
     * @maxLength 100
     */
  timeZone: string;
}

export interface WorkoutSyncResult {
  synced: number;
  score: number;
  scoreGained: number;
  rank: string;
  rankLevel: number;
  rankedUp: boolean;
}

export interface LeaderboardEntry {
  userId: string;
  username: string;
  displayName: string;
  score: number;
  rank: string;
  rankLevel: number;
  isYou: boolean;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  yourPosition: number;
}

export interface FriendSummary {
  userId: string;
  username: string;
  displayName: string;
  score: number;
  rank: string;
  rankLevel: number;
  sharingEnabled: boolean;
}

export interface FriendRequestSummary {
  id: number;
  userId: string;
  username: string;
  displayName: string;
  createdAt: string;
}

export interface FriendsResponse {
  friends: FriendSummary[];
  incoming: FriendRequestSummary[];
  outgoing: FriendRequestSummary[];
}

export interface FriendRequestInput {
  /**
     * @minLength 3
     * @maxLength 20
     */
  username: string;
}

export type FriendRequestDecisionDecision = typeof FriendRequestDecisionDecision[keyof typeof FriendRequestDecisionDecision];


export const FriendRequestDecisionDecision = {
  accept: 'accept',
  decline: 'decline',
  cancel: 'cancel',
} as const;

export interface FriendRequestDecision {
  decision: FriendRequestDecisionDecision;
}

export type FriendRequestDecisionResultStatus = typeof FriendRequestDecisionResultStatus[keyof typeof FriendRequestDecisionResultStatus];


export const FriendRequestDecisionResultStatus = {
  accepted: 'accepted',
  declined: 'declined',
  cancelled: 'cancelled',
} as const;

export interface FriendRequestDecisionResult {
  status: FriendRequestDecisionResultStatus;
}

export type GetCommunityMeParams = {
/**
 * @maxLength 100
 */
timeZone?: string;
};