# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: community
    description: ShadowFit profiles, friends, and scoreboards
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /community/me:
    get:
      operationId: getCommunityMe
      tags: [community]
      summary: Get or create the signed-in user's ShadowFit profile
      parameters:
        - name: timeZone
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: Current profile and rank
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityProfile"
    put:
      operationId: updateCommunityMe
      tags: [community]
      summary: Update the current user's public profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityProfile"
        "409":
          description: Username is already in use
  /community/sharing:
    put:
      operationId: setCommunitySharing
      tags: [community]
      summary: Enable or pause Shadow Score sharing with accepted friends
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SharingPreferenceInput"
      responses:
        "200":
          description: Updated score sharing preference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharingPreferenceResult"
  /community/workouts/sync:
    post:
      operationId: syncCommunityWorkouts
      tags: [community]
      summary: Sync locally recorded workout sessions for the signed-in user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutSyncInput"
      responses:
        "200":
          description: Updated score and rank
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutSyncResult"
  /community/leaderboard:
    get:
      operationId: getCommunityLeaderboard
      tags: [community]
      summary: Get the signed-in user and accepted friends ranked by Shadow Score
      responses:
        "200":
          description: Friends-only leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
  /community/friends:
    get:
      operationId: getCommunityFriends
      tags: [community]
      summary: List accepted friends and pending friend requests
      responses:
        "200":
          description: Friends and requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendsResponse"
  /community/friend-requests:
    post:
      operationId: createFriendRequest
      tags: [community]
      summary: Send a friend request by username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequestInput"
      responses:
        "201":
          description: Friend request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequestSummary"
        "404":
          description: No profile matches that username
        "409":
          description: A request or friendship already exists
  /community/friend-requests/{requestId}:
    patch:
      operationId: respondFriendRequest
      tags: [community]
      summary: Accept, decline, or cancel a friend request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequestDecision"
      responses:
        "200":
          description: Request updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequestDecisionResult"
        "404":
          description: Request not found
        "403":
          description: Request cannot be changed by this user
  /community/friends/{friendUserId}:
    delete:
      operationId: removeCommunityFriend
      tags: [community]
      summary: Remove an accepted friend
      parameters:
        - name: friendUserId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Friendship removed
        "404":
          description: Friendship not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ProfileInput:
      type: object
      required: [username, displayName]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
          pattern: "^[a-z0-9_]+$"
        displayName:
          type: string
          minLength: 1
          maxLength: 32
    CommunityProfile:
      type: object
      required: [userId, username, displayName, score, rank, rankLevel, friendCount, nextRankScore, sharingEnabled]
      properties:
        userId: { type: string }
        username: { type: string }
        displayName: { type: string }
        score: { type: integer }
        rank: { type: string }
        rankLevel: { type: integer }
        friendCount: { type: integer }
        nextRankScore: { type: ["integer", "null"] }
        sharingEnabled: { type: boolean }
    SharingPreferenceInput:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }
    SharingPreferenceResult:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }
    WorkoutSessionInput:
      type: object
      required: [clientSessionId, workoutId, sets, minutes, completedAt, localDay]
      properties:
        clientSessionId: { type: string, minLength: 1, maxLength: 80 }
        workoutId: { type: string, minLength: 1, maxLength: 100 }
        sets: { type: integer, minimum: 1, maximum: 100 }
        minutes: { type: integer, minimum: 1, maximum: 240 }
        completedAt: { type: string, format: date-time }
        localDay: { type: string, pattern: "^\\d{4}-\\d{2}-\\d{2}$" }
    WorkoutSyncInput:
      type: object
      required: [sessions, timeZone]
      properties:
        sessions:
          type: array
          minItems: 1
          maxItems: 250
          items:
            $ref: "#/components/schemas/WorkoutSessionInput"
        timeZone: { type: string, minLength: 1, maxLength: 100 }
    WorkoutSyncResult:
      type: object
      required: [synced, score, scoreGained, rank, rankLevel, rankedUp]
      properties:
        synced: { type: integer }
        score: { type: integer }
        scoreGained: { type: integer }
        rank: { type: string }
        rankLevel: { type: integer }
        rankedUp: { type: boolean }
    LeaderboardEntry:
      type: object
      required: [userId, username, displayName, score, rank, rankLevel, isYou]
      properties:
        userId: { type: string }
        username: { type: string }
        displayName: { type: string }
        score: { type: integer }
        rank: { type: string }
        rankLevel: { type: integer }
        isYou: { type: boolean }
    LeaderboardResponse:
      type: object
      required: [entries, yourPosition]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
        yourPosition: { type: integer }
    FriendSummary:
      type: object
      required: [userId, username, displayName, score, rank, rankLevel, sharingEnabled]
      properties:
        userId: { type: string }
        username: { type: string }
        displayName: { type: string }
        score: { type: integer }
        rank: { type: string }
        rankLevel: { type: integer }
        sharingEnabled: { type: boolean }
    FriendRequestSummary:
      type: object
      required: [id, userId, username, displayName, createdAt]
      properties:
        id: { type: integer }
        userId: { type: string }
        username: { type: string }
        displayName: { type: string }
        createdAt: { type: string, format: date-time }
    FriendsResponse:
      type: object
      required: [friends, incoming, outgoing]
      properties:
        friends:
          type: array
          items:
            $ref: "#/components/schemas/FriendSummary"
        incoming:
          type: array
          items:
            $ref: "#/components/schemas/FriendRequestSummary"
        outgoing:
          type: array
          items:
            $ref: "#/components/schemas/FriendRequestSummary"
    FriendRequestInput:
      type: object
      required: [username]
      properties:
        username: { type: string, minLength: 3, maxLength: 20 }
    FriendRequestDecision:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [accept, decline, cancel]
    FriendRequestDecisionResult:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, declined, cancelled]