# lib/api-zod/src/generated/types/communityProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CommunityProfile {
  userId: string;
  username: string;
  displayName: string;
  score: number;
  rank: string;
  rankLevel: number;
  friendCount: number;
  /** @nullable */
  nextRankScore: number | null;
  sharingEnabled: boolean;
}