# lib/api-zod/src/generated/types/friendRequestDecision.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FriendRequestDecisionDecision } from './friendRequestDecisionDecision';

export interface FriendRequestDecision {
  decision: FriendRequestDecisionDecision;
}