# lib/api-zod/src/generated/types/friendRequestDecisionResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FriendRequestDecisionResultStatus } from './friendRequestDecisionResultStatus';

export interface FriendRequestDecisionResult {
  status: FriendRequestDecisionResultStatus;
}