# lib/api-zod/src/generated/types/friendRequestInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FriendRequestInput {
  /**
     * @minLength 3
     * @maxLength 20
     */
  username: string;
}