# lib/api-zod/src/generated/types/friendRequestSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FriendRequestSummary {
  id: number;
  userId: string;
  username: string;
  displayName: string;
  createdAt: Date;
}