# lib/api-zod/src/generated/types/friendSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FriendSummary {
  userId: string;
  username: string;
  displayName: string;
  score: number;
  rank: string;
  rankLevel: number;
  sharingEnabled: boolean;
}