# lib/api-zod/src/generated/types/friendsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FriendRequestSummary } from './friendRequestSummary';
import type { FriendSummary } from './friendSummary';

export interface FriendsResponse {
  friends: FriendSummary[];
  incoming: FriendRequestSummary[];
  outgoing: FriendRequestSummary[];
}