# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  userId: string;
  username: string;
  displayName: string;
  score: number;
  rank: string;
  rankLevel: number;
  isYou: boolean;
}