# lib/api-zod/src/generated/types/profileInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileInput {
  /**
     * @minLength 3
     * @maxLength 20
     * @pattern ^[a-z0-9_]+$
     */
  username: string;
  /**
     * @minLength 1
     * @maxLength 32
     */
  displayName: string;
}