# lib/api-zod/src/generated/types/workoutSyncInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutSessionInput } from './workoutSessionInput';

export interface WorkoutSyncInput {
  /**
     * @minItems 1
     * @maxItems 250
     */
  sessions: WorkoutSessionInput[];
  /**
     * @minLength 1
     * @maxLength 100
     */
  timeZone: string;
}